Latest commit

+++ controllers/courses.js
@@ -9,7 +9,7 @@ const asyncHandler = require("../middleware/async");
 // @access  Public
 exports.getCourses = asyncHandler(async (req, res, next) => {
   if (req.params.bootcampId) {
-    const course = await Course.find({ bootcamp: req.params.bootcampId });
+    const courses = await Course.find({ bootcamp: req.params.bootcampId });
 
     return res.status(200).json({
       success: true,
@@ -20,7 +20,6 @@ exports.getCourses = asyncHandler(async (req, res, next) => {
     res.status(200).json(res.advancedResults);
   }
 });
-
 // @desc    Get single course
 // @route   GET /api/v1/courses/:id
 // @access  Public
@@ -46,6 +45,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // POST /api/v1/bootcamps/:bootcampId/courses
 // @access  Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
+  console.log("Authenticated User ID:", req.user.id);
   const bootcampId = req.params.bootcampId;
   req.body.bootcamp = bootcampId;
   req.body.user = req.user.id;
@@ -62,7 +62,7 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
   if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
     return next(
       new ErrorResponse(
-        `User ${req.params.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
+        `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`,
         401
       )
     );
@@ -88,7 +88,16 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
       404
     );
   }
-
+  console.log(req.user);
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.name} is not authorized to update course ${course.title}`,
+        401
+      )
+    );
+  }
   course = await Course.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
     runValidators: true,
@@ -113,7 +122,22 @@ exports.deleteCourse = asyncHandler(async (req, res, next) => {
     );
   }
 
+  // Make sure user is course owner
+  if (course.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete course ${course._id}`,
+        401
+      )
+    );
+  }
+
+  const bootcampId = course.bootcamp;
+
   await course.deleteOne();
+
+  // Manually call getAverageCost
+  await Course.getAverageCost(bootcampId);
 
   res.status(200).json({
     success: true,
